# sha2: wasm32 simd128 backend for SHA-512