--- a/.github/workflows/sha2.yml
+++ b/.github/workflows/sha2.yml
@@ -196,6 +196,38 @@ jobs:
         env:
           RUSTFLAGS: -Dwarnings --cfg sha2_backend="riscv-zknh-compact" -C target-feature=+zknh,+zbkb
 
+  # wasmtime tests
+  wasm:
+    needs: set-msrv
+    strategy:
+      matrix:
+        include:
+          # without simd
+          - rust: ${{needs.set-msrv.outputs.msrv}}
+            flags: "-C target-feature=-simd128"
+          - rust: stable
+            flags: "-C target-feature=-simd128"
+
+          # with simd
+          - rust: ${{needs.set-msrv.outputs.msrv}}
+            flags: "-C target-feature=+simd128"
+          - rust: stable
+            flags: "-C target-feature=+simd128"
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: RustCrypto/actions/cargo-cache@master
+      - uses: dtolnay/rust-toolchain@master
+        with:
+          toolchain: ${{ matrix.rust }}
+          targets: wasm32-wasi
+      - uses: RustCrypto/actions/cargo-hack-install@master
+      - uses: jcbhmr/setup-wasmtime@v2
+      - run: cargo hack test --feature-powerset --target wasm32-wasi
+        env:
+          RUSTFLAGS: ${{ matrix.flags }}
+          CARGO_TARGET_WASM32_WASI_RUNNER: wasmtime
+  
   minimal-versions:
     uses: RustCrypto/actions/.github/workflows/minimal-versions.yml@master
     with: