--- a/.github/workflows/sha2.yml
+++ b/.github/workflows/sha2.yml
@@ -205,14 +205,18 @@ jobs:
           # without simd
           - rust: ${{needs.set-msrv.outputs.msrv}}
             flags: "-C target-feature=-simd128"
+            target_name: wasm32-wasi # Changed in Rust >=1.81
           - rust: stable
             flags: "-C target-feature=-simd128"
+            target_name: wasm32-wasip1
 
           # with simd
           - rust: ${{needs.set-msrv.outputs.msrv}}
             flags: "-C target-feature=+simd128"
+            target_name: wasm32-wasi
           - rust: stable
             flags: "-C target-feature=+simd128"
+            target_name: wasm32-wasip1
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
@@ -220,14 +224,15 @@ jobs:
       - uses: dtolnay/rust-toolchain@master
         with:
           toolchain: ${{ matrix.rust }}
-          targets: wasm32-wasi
+          targets: ${{ matrix.target_name }}
       - uses: RustCrypto/actions/cargo-hack-install@master
       - uses: jcbhmr/setup-wasmtime@v2
-      - run: cargo hack test --feature-powerset --target wasm32-wasi
+      - run: cargo hack test --feature-powerset --target ${{ matrix.target_name }}
         env:
           RUSTFLAGS: ${{ matrix.flags }}
           CARGO_TARGET_WASM32_WASI_RUNNER: wasmtime
-  
+          CARGO_TARGET_WASM32_WASIP1_RUNNER: wasmtime
+
   minimal-versions:
     uses: RustCrypto/actions/.github/workflows/minimal-versions.yml@master
     with: