--- a/.github/workflows/sha2.yml
+++ b/.github/workflows/sha2.yml
@@ -21,7 +21,7 @@ jobs:
   set-msrv:
     uses: RustCrypto/actions/.github/workflows/set-msrv.yml@master
     with:
-      msrv: 1.79.0
+      msrv: 1.81.0
 
   # Builds for no_std platforms
   build:
@@ -205,18 +205,14 @@ jobs:
           # without simd
           - rust: ${{needs.set-msrv.outputs.msrv}}
             flags: "-C target-feature=-simd128"
-            target_name: wasm32-wasi # Changed in Rust >=1.81
           - rust: stable
             flags: "-C target-feature=-simd128"
-            target_name: wasm32-wasip1
 
           # with simd
           - rust: ${{needs.set-msrv.outputs.msrv}}
             flags: "-C target-feature=+simd128"
-            target_name: wasm32-wasi
           - rust: stable
             flags: "-C target-feature=+simd128"
-            target_name: wasm32-wasip1
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
@@ -224,13 +220,12 @@ jobs:
       - uses: dtolnay/rust-toolchain@master
         with:
           toolchain: ${{ matrix.rust }}
-          targets: ${{ matrix.target_name }}
+          targets: wasm32-wasip1
       - uses: RustCrypto/actions/cargo-hack-install@master
       - uses: jcbhmr/setup-wasmtime@v2
-      - run: cargo hack test --feature-powerset --target ${{ matrix.target_name }}
+      - run: cargo hack test --feature-powerset --target wasm32-wasip1
         env:
           RUSTFLAGS: ${{ matrix.flags }}
-          CARGO_TARGET_WASM32_WASI_RUNNER: wasmtime
           CARGO_TARGET_WASM32_WASIP1_RUNNER: wasmtime
 
   minimal-versions: