--- a/sha2/src/sha512.rs
+++ b/sha2/src/sha512.rs
@@ -27,6 +27,9 @@ cfg_if::cfg_if! {
     } else if #[cfg(target_arch = "loongarch64")] {
         mod loongarch64_asm;
         use loongarch64_asm::compress;
+    } else if #[cfg(target_arch = "wasm32")] {
+        mod wasm32;
+        use wasm32::compress;
     } else {
         mod soft;
         use soft::compress;