--- a/sha2/src/sha512.rs
+++ b/sha2/src/sha512.rs
@@ -27,7 +27,7 @@ cfg_if::cfg_if! {
     } else if #[cfg(target_arch = "loongarch64")] {
         mod loongarch64_asm;
         use loongarch64_asm::compress;
-    } else if #[cfg(target_arch = "wasm32")] {
+    } else if #[cfg(all(target_arch = "wasm32", target_feature = "simd128"))] {
         mod wasm32;
         use wasm32::compress;
     } else {