--- a/sha2/src/sha512/wasm32.rs
+++ b/sha2/src/sha512/wasm32.rs
@@ -5,7 +5,6 @@ use core::mem::size_of;
 
 use crate::consts::K64;
 
-#[target_feature(enable = "simd128")]
 pub fn compress(state: &mut [u64; 8], blocks: &[[u8; 128]]) {
     let mut ms = [u64x2(0, 0); 8];
     let mut x = [u64x2(0, 0); 8];