@newpavlov 2024-11-03
@max-te 
I think we are good to merge. But could you measure performance of the software backend with enabled `simd128` target feature on the same hardware? You can do it with this command:
```
RUSTFLAGS='--cfg sha2_backend="soft" -C target-feature=+simd128' cargo +nightly bench --target wasm32-wasi
```
I would like to add these results to the merge commit message.