@max-te 2024-11-03
Sure, I added that benchmark to the PR description and updated the other ones.